requirements.txt
flask
python-dotenv
requests
pyodbc
sqlalchemy
flask-cors
gunicorn
